Move logic into api module and create list_milestones method
