+++ tests/test_clubhouse.py
@@ -2,4 +2,4 @@ from clubhouse import __version__
 
 
 def test_version():
-    assert __version__ == "0.1.0"
+    assert __version__ == "0.3.0"
